language-proficiency-backend/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
cachetools==5.3.3
certifi==2024.7.4
charset-normalizer==2.0.12
click==8.1.7
distro==1.9.0
Flask==2.1.0
Flask-Cors==3.0.10
google-api-core==2.19.1
google-auth==2.31.0
google-cloud-speech==2.13.1
googleapis-common-protos==1.63.2
grpcio==1.64.1
grpcio-status==1.64.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==2.0.0
openai==1.35.10
pandas==2.2.2
proto-plus==1.24.0
protobuf==5.27.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-dotenv==0.19.2
pytz==2024.1
requests==2.27.1
rsa==4.9
six==1.16.0
sniffio==1.3.1
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.19
Werkzeug==3.0.3
